requirements.txt
networkx==2.4
scikit-learn==0.24.0
stellargraph==1.2.1
spacy==3.0.3
pandas==1.1.3
numpy==1.19.2
node2vec==0.3.3
Keras==2.0.2
tensorflow==2.4.1
communities==2.2.0
gensim==3.8.3
matplotlib==3.3.4
nltk==3.5
transformers==4.17.0
neo4j-driver==4.4.2
chardet==4.0.0
dask==2021.3.0
fsspec>=0.3.3
ray==2.3.1
beautifulsoup4==4.12.2
